+++ src/sputnik/src/js/apis/node/llrt/mod.rs
@@ -2,6 +2,8 @@ mod llrt_buffer;
 #[rustfmt::skip]
 mod llrt_utils;
 mod blob;
+mod llrt_url;
 mod url;
 
 pub use blob::*;
+pub use url::*;
